Unify profile and identity && topbar styles
